+++ genium/src/test/resources/smoke/nesting/output/android/jni/StructConversion.h
@@ -4,7 +4,9 @@
 #include "JniBase.h"
 #include "JniCppConversionUtils.h"
 #include "com_example_smoke_FreePoint.h"
+#include "com_example_smoke_NestedReferences.h"
 #include "smoke/FreePoint.h"
+#include "smoke/NestedReferences.h"
 #include "smoke/UnusedTopLevelPoint.h"
 #include <jni.h>
 #include <memory>
@@ -17,6 +19,10 @@ namespace jni
 ::genium::optional<::smoke::FreePoint> convert_from_jni(JNIEnv* _jenv, const JniReference<jobject>& _jinput, ::genium::optional<::smoke::FreePoint>* dummy);
 JniReference<jobject> convert_to_jni(JNIEnv* _jenv, const ::smoke::FreePoint& _ninput);
 JniReference<jobject> convert_to_jni(JNIEnv* _jenv, const ::genium::optional<::smoke::FreePoint> _ninput);
+::smoke::NestedReferences::NestedReferences convert_from_jni(JNIEnv* _jenv, const JniReference<jobject>& _jinput, ::smoke::NestedReferences::NestedReferences* dummy);
+::genium::optional<::smoke::NestedReferences::NestedReferences> convert_from_jni(JNIEnv* _jenv, const JniReference<jobject>& _jinput, ::genium::optional<::smoke::NestedReferences::NestedReferences>* dummy);
+JniReference<jobject> convert_to_jni(JNIEnv* _jenv, const ::smoke::NestedReferences::NestedReferences& _ninput);
+JniReference<jobject> convert_to_jni(JNIEnv* _jenv, const ::genium::optional<::smoke::NestedReferences::NestedReferences> _ninput);
 ::smoke::UnusedTopLevelPoint convert_from_jni(JNIEnv* _jenv, const JniReference<jobject>& _jinput, ::smoke::UnusedTopLevelPoint* dummy);
 ::genium::optional<::smoke::UnusedTopLevelPoint> convert_from_jni(JNIEnv* _jenv, const JniReference<jobject>& _jinput, ::genium::optional<::smoke::UnusedTopLevelPoint>* dummy);
 JniReference<jobject> convert_to_jni(JNIEnv* _jenv, const ::smoke::UnusedTopLevelPoint& _ninput);