+++ genium/src/test/resources/smoke/nesting/output/cpp/include/smoke/NestedReferences.h
@@ -1,0 +1,32 @@
+// -------------------------------------------------------------------------------------------------
+//
+//
+// -------------------------------------------------------------------------------------------------
+#pragma once
+#include "genium/Export.h"
+#include <memory>
+#include <string>
+namespace smoke {
+    class NestedReferences;
+}
+namespace smoke {
+class _GENIUM_CPP_EXPORT NestedReferences {
+public:
+    NestedReferences();
+    virtual ~NestedReferences() = 0;
+public:
+    struct _GENIUM_CPP_EXPORT NestedReferences {
+        ::std::string string_field;
+        NestedReferences( );
+        NestedReferences( const ::std::string& string_field );
+    };
+public:
+    /**
+     *
+     * \param[in] struct1
+     * \param[in] struct2
+     * \return @NotNull
+     */
+    virtual ::std::shared_ptr< ::smoke::NestedReferences > inside_out( const ::smoke::NestedReferences::NestedReferences& struct1, const ::smoke::NestedReferences::NestedReferences& struct2 ) = 0;
+};
+}