+++ lime-runtime/src/main/java/com/here/genium/model/lime/LimeAmbiguousEnumeratorRef.kt
@@ -21,13 +21,7 @@ package com.here.genium.model.lime
 
 /**
  * An ambiguous reference to an enumerator element, represented by [relativePath]. The ambiguity is
- * resolved with the following logic:
- * * for each import in [imports], try to match the last component of the import to the first
- * component of the [relativePath]. If it matches, try to resolve into to an actual element through
- * the [referenceMap].
- * * if import resolution fails, try appending [relativePath] to each parent path in [parentPaths]
- * and try resolving the combined path to an actual element through the [referenceMap].
- * * if both approaches fail, the reference is invalid and a [LimeModelLoaderException] is thrown.
+ * resolved through the [LimeAmbiguityResolver].
  *
  * The resolution logic is "lazy": if it succeed on the first call then the result is stored and the
  * stored result is used on subsequent calls instead.
@@ -42,14 +36,6 @@ class LimeAmbiguousEnumeratorRef(
     override val elementFullName by lazy { enumerator.path.toString() }
 
     override val enumerator by lazy {
-        for (limePath in imports.filter { it.name == relativePath.first() }) {
-            val key = limePath.child(relativePath.drop(1)).toString()
-            return@lazy referenceMap[key] as? LimeEnumerator ?: continue
-        }
-        for (limePath in parentPaths) {
-            val key = limePath.child(relativePath).toString()
-            return@lazy referenceMap[key] as? LimeEnumerator ?: continue
-        }
-        throw LimeModelLoaderException("Enumerator $relativePath was not found")
+        LimeAmbiguityResolver.resolve<LimeEnumerator>(relativePath, parentPaths, imports, referenceMap)
     }
 }